+++ desmume/src/lua-engine.h
@@ -20,11 +20,11 @@
 
 #include "types.h"
 
-void OpenLuaContext(int uid, void(*print)(int uid, const char* str) = 0, void(*onstart)(int uid) = 0, void(*onstop)(int uid, bool statusOK) = 0);
-void RunLuaScriptFile(int uid, const char* filename);
-void StopLuaScript(int uid);
-void RequestAbortLuaScript(int uid, const char* message = 0);
-void CloseLuaContext(int uid);
+void OpenLuaContext(uintptr_t uid, void(*print)(uintptr_t uid, const char* str) = 0, void(*onstart)(uintptr_t uid) = 0, void(*onstop)(uintptr_t uid, bool statusOK) = 0);
+void RunLuaScriptFile(uintptr_t uid, const char* filename);
+void StopLuaScript(uintptr_t uid);
+void RequestAbortLuaScript(uintptr_t uid, const char* message = 0);
+void CloseLuaContext(uintptr_t uid);
 bool AnyLuaActive();
 
 enum LuaCallID
@@ -89,9 +89,9 @@ struct LuaSaveData
 
 	Record* recordList;
 
-	void SaveRecord(int uid, unsigned int key); // saves Lua stack into a record and pops it
-	void LoadRecord(int uid, unsigned int key, unsigned int itemsToLoad) const; // pushes a record's data onto the Lua stack
-	void SaveRecordPartial(int uid, unsigned int key, int idx); // saves part of the Lua stack (at the given index) into a record and does NOT pop anything
+	void SaveRecord(uintptr_t uid, unsigned int key); // saves Lua stack into a record and pops it
+	void LoadRecord(uintptr_t uid, unsigned int key, unsigned int itemsToLoad) const; // pushes a record's data onto the Lua stack
+	void SaveRecordPartial(uintptr_t uid, unsigned int key, int idx); // saves part of the Lua stack (at the given index) into a record and does NOT pop anything
 
 	void ExportRecords(void* file) const; // writes all records to an already-open file
 	void ImportRecords(void* file); // reads records from an already-open file
